requirements.txt
telethon==1.34.0
aiosqlite==0.19.0
PyYAML==6.0.1
aiohttp==3.9.3
